requirements.txt
discord.py
requests
watchdog
psutil